+++ packages/offline-browser-sdk/src/socket-io-events/game-ended.ts
@@ -1,0 +1,13 @@
+import { GameState, GameStateSchema } from "gbraver-burst-core";
+import { z } from "zod";
+
+/** socket.ioイベントgameEnded */
+export type GameEnded = {
+  /** 更新されたゲームステートの履歴 */
+  updatedStateHistory: GameState[];
+};
+
+/** socket.ioイベントgameEndedのZodスキーマ */
+export const GameEndedSchema = z.object({
+  updatedStateHistory: z.array(GameStateSchema),
+});