+++ packages/offline-stub/src/index.ts
@@ -1,0 +1,217 @@
+import { createOfflineBrowserSDK } from "@gbraver-burst-network/offline-browser-sdk";
+import { OfflineBattleSDK } from "@gbraver-burst-network/offline-browser-sdk/lib/offline-battle-sdk";
+import {
+  Armdozers,
+  Command,
+  CommandSchema,
+  Pilots,
+  Selectable,
+} from "gbraver-burst-core";
+
+/** バックエンドURL */
+const BACKEND_URL = process.env.BACKEND_URL || "http://localhost:3000";
+
+/**
+ * アームドーザのセレクターを取得する
+ * @returns アームドーザのセレクター
+ */
+const getArmdozerSelector = () =>
+  document.getElementById("armdozer") as HTMLSelectElement;
+
+/**
+ * アームドーザの選択肢を更新する
+ */
+const updateArmdozerOptions = () => {
+  const armdozerSelect = getArmdozerSelector();
+  Armdozers.forEach((armdozer) => {
+    const option = document.createElement("option");
+    option.value = armdozer.id;
+    option.text = armdozer.name;
+    armdozerSelect.appendChild(option);
+  });
+};
+
+/**
+ * パイロットのセレクターを取得する
+ * @returns パイロットのセレクター
+ */
+const getPilotSelector = () =>
+  document.getElementById("pilot") as HTMLSelectElement;
+
+/**
+ * パイロットの選択肢を更新する
+ */
+const updatePilotOptions = () => {
+  const pilotSelect = getPilotSelector();
+  Pilots.forEach((pilot) => {
+    const option = document.createElement("option");
+    option.value = pilot.id;
+    option.text = pilot.name;
+    pilotSelect.appendChild(option);
+  });
+};
+
+/**
+ * キャラクターセレクターを操作可能にする
+ */
+const enableSelector = () => {
+  document.getElementById("armdozer")?.removeAttribute("disabled");
+  document.getElementById("pilot")?.removeAttribute("disabled");
+  document.getElementById("enter-room")?.removeAttribute("disabled");
+};
+
+/**
+ * キャラクターセレクターを操作不可能にする
+ */
+const disabledSelector = () => {
+  document.getElementById("armdozer")?.setAttribute("disabled", "true");
+  document.getElementById("pilot")?.setAttribute("disabled", "true");
+  document.getElementById("enter-room")?.setAttribute("disabled", "true");
+};
+
+/**
+ * コマンドセレクターを非表示にする
+ */
+const hiddenCommands = () => {
+  document.getElementById("commands")?.setAttribute("style", "display:none");
+};
+
+/**
+ * コマンドセレクターを表示する
+ */
+const showCommands = () => {
+  document.getElementById("commands")?.setAttribute("style", "display:block");
+};
+
+/**
+ * 選択されているコマンドを取得する
+ * @returns 選択されているコマンド
+ */
+const getSelectedCommand = (): Command => {
+  const commandSelector = document.getElementById(
+    "command-selector",
+  ) as HTMLSelectElement;
+  if (!commandSelector) {
+    throw new Error("Command selector not found");
+  }
+
+  const command = JSON.parse(commandSelector.value);
+  const parsedCommand = CommandSchema.parse(command);
+  return parsedCommand;
+};
+
+/**
+ * コマンドセレクターを更新する
+ * @param selectable 選択可能なコマンド
+ */
+const updateCommands = (selectable: Selectable) => {
+  const commandSelector = document.getElementById(
+    "command-selector",
+  ) as HTMLSelectElement;
+  commandSelector.innerHTML = "";
+  selectable.command.forEach((command) => {
+    const option = document.createElement("option");
+    const jsonStr = JSON.stringify(command);
+    option.value = jsonStr;
+    option.text = jsonStr;
+    commandSelector.appendChild(option);
+  });
+};
+
+/**
+ * 現在のバトルSDK
+ * 対戦以外ではnullが入る
+ */
+let battleSDK: OfflineBattleSDK | null = null;
+
+/** スタブのエントリポイント */
+window.onload = () => {
+  hiddenCommands();
+  updateArmdozerOptions();
+  updatePilotOptions();
+  const sdk = createOfflineBrowserSDK({ backendURL: BACKEND_URL });
+
+  /**
+   * 入室ボタンのクリックイベント
+   */
+  document.getElementById("enter-room")?.addEventListener("click", async () => {
+    const armdozerSelect = getArmdozerSelector();
+    const pilotSelect = getPilotSelector();
+    disabledSelector();
+
+    battleSDK = await sdk.enterRoom({
+      armdozerId: armdozerSelect.value,
+      pilotId: pilotSelect.value,
+    });
+
+    console.log("matched", battleSDK);
+    const inputCommand = battleSDK.initialState.findLast(
+      (s) => s.effect.name === "InputCommand",
+    );
+    if (inputCommand?.effect.name !== "InputCommand") {
+      return;
+    }
+
+    const playerId = battleSDK.player.playerId;
+    const commands = inputCommand.effect.players.find(
+      (p) => p.playerId === playerId,
+    );
+    if (commands?.selectable) {
+      updateCommands(commands);
+      showCommands();
+    }
+  });
+
+  /**
+   * コマンド送信ボタンのクリックイベント
+   */
+  document
+    .getElementById("send-command")
+    ?.addEventListener("click", async () => {
+      hiddenCommands();
+      if (!battleSDK) {
+        return;
+      }
+
+      let lastCommand = getSelectedCommand();
+      const maxPollingAttempts = 100;
+      for (let i = 0; i < maxPollingAttempts; i++) {
+        const updatedStateHistory = await battleSDK.progress(lastCommand);
+        console.log("game progressed", updatedStateHistory);
+        const lastState = updatedStateHistory.at(-1);
+
+        if (lastState?.effect.name === "GameEnd") {
+          hiddenCommands();
+          enableSelector();
+          sdk.closeConnection();
+          return;
+        }
+
+        if (lastState?.effect.name !== "InputCommand") {
+          return;
+        }
+
+        const commands = lastState.effect.players.find(
+          (p) => p.playerId === battleSDK?.player.playerId,
+        );
+        if (!commands) {
+          return;
+        }
+
+        if (commands.selectable) {
+          updateCommands(commands);
+          showCommands();
+          return;
+        }
+
+        lastCommand = commands.nextTurnCommand;
+      }
+    });
+
+  /**
+   * エラーが発生した時の処理
+   */
+  sdk.notifyError().subscribe((error) => {
+    console.error("error occurred", error);
+  });
+};