+++ packages/backend-app/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gbraver-burst-network/backend-app
 
+## 1.19.0
+
+### Minor Changes
+
+- バトルSDKのインタフェースを統一
+
+## 1.18.0
+
+### Minor Changes
+
+- オフライン用SDKを追加した
+
 ## 1.17.2
 
 ### Patch Changes