+++ packages/serverless-stub/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @gbraver-burst-network/serverless-stub
 
+## 1.19.0
+
+### Minor Changes
+
+- バトルSDKのインタフェースを統一
+
+### Patch Changes
+
+- Updated dependencies
+  - @gbraver-burst-network/browser-sdk@1.19.0
+
+## 1.18.0
+
+### Minor Changes
+
+- オフライン用SDKを追加した
+
+### Patch Changes
+
+- Updated dependencies
+  - @gbraver-burst-network/browser-sdk@1.18.0
+
 ## 1.17.2
 
 ### Patch Changes